! 2012-02-21 - Chris Jones  - Update the way some references to Particles and Vertices are stored, to    use 64 bit intergers, to avoid packing problems.
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@135639 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 293b34924224a81a3fb3c6f1877ff08cb1bd7778

+++ LHCb/Event/PackedEvent/Event/PackedParticle.h
@@ -51,10 +51,10 @@ namespace LHCb
         vertex(-1),
         proto(-1),
         firstDaughter(0), lastDaughter(0)
-    {}
+    { }
 
     // packed data members
-    int key;               ///< referenceLong to the original container + key of the particle
+    long long key;         ///< referenceLong to the original container + key of the particle
     int particleID;        ///< PID Code
     int measMass;          ///< Measured mass
     int measMassErr;       ///< Error on the measured mass
@@ -83,13 +83,13 @@ namespace LHCb
     unsigned short int firstExtra, lastExtra;
 
     // End Vertex
-    int vertex;  
+    long long vertex;  
 
     // ProtoParticle
     int proto;
 
     // daughters
-    unsigned short int firstDaughter, lastDaughter;  
+    unsigned int firstDaughter, lastDaughter;  
 
   };
 
@@ -126,7 +126,7 @@ namespace LHCb
     typedef std::vector<PackedExtraInfo> PackedExtraInfoVector;
 
     /// Daughters
-    typedef std::vector<int> Daughters;
+    typedef std::vector<long long> Daughters;
 
   public:
